+++ sdk/core/azure-core-amqp/src/main/java/com/azure/core/amqp/implementation/ReactorSession.java
@@ -55,6 +55,7 @@ import java.util.function.Consumer;
 import static com.azure.core.amqp.implementation.AmqpLoggingUtils.addErrorCondition;
 import static com.azure.core.amqp.implementation.AmqpLoggingUtils.addSignalTypeAndResult;
 import static com.azure.core.amqp.implementation.AmqpLoggingUtils.createContextWithConnectionId;
+import static com.azure.core.amqp.implementation.ClientConstants.CLIENT_ID_KEY;
 import static com.azure.core.amqp.implementation.ClientConstants.ENTITY_PATH_KEY;
 import static com.azure.core.amqp.implementation.ClientConstants.LINK_NAME_KEY;
 import static com.azure.core.amqp.implementation.ClientConstants.NOT_APPLICABLE;
@@ -96,6 +97,9 @@ public class ReactorSession implements AmqpSession {
 
     private final AtomicReference<TransactionCoordinator> transactionCoordinator = new AtomicReference<>();
     private final Flux<AmqpShutdownSignal> shutdownSignals;
+
+    public static final Symbol CLIENT_ID = Symbol.getSymbol(AmqpConstants.VENDOR + ":receiver-name");
+
 
     /**
      * Creates a new AMQP session using proton-j.
@@ -525,13 +529,16 @@ public class ReactorSession implements AmqpSession {
 
         final Sender sender = session.sender(linkName);
         sender.setTarget(target);
-
-        final Source source = new Source();
-        sender.setSource(source);
         sender.setSenderSettleMode(SenderSettleMode.UNSETTLED);
 
         if (linkProperties != null && linkProperties.size() > 0) {
             sender.setProperties(linkProperties);
+            String clientId = (String) linkProperties.get(CLIENT_ID);
+            if (clientId != null && !clientId.equals("")) {
+                final Source source = new Source();
+                source.setAddress(clientId);
+                sender.setSource(source);
+            }
         }
 
         final SendLinkHandler sendLinkHandler = handlerProvider.createSendLinkHandler(
@@ -583,15 +590,18 @@ public class ReactorSession implements AmqpSession {
 
         receiver.setSource(source);
 
-        final Target target = new Target();
-        receiver.setTarget(target);
-
         // Use explicit settlement via dispositions (not pre-settled)
         receiver.setSenderSettleMode(senderSettleMode);
         receiver.setReceiverSettleMode(receiverSettleMode);
 
         if (receiverProperties != null && !receiverProperties.isEmpty()) {
             receiver.setProperties(receiverProperties);
+            String clientId = (String) receiverProperties.get(CLIENT_ID);
+            if (clientId != null && clientId != "") {
+                final Target target = new Target();
+                target.setAddress(clientId);
+                receiver.setTarget(target);
+            }
         }
 
         if (receiverDesiredCapabilities != null && receiverDesiredCapabilities.length > 0) {